backend/requirements.txt
# FastAPI 后端依赖
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# CORS 支持（FastAPI 内置，无需额外安装）

# 项目核心依赖（从根目录 requirements.txt 继承）
neo4j>=5.14.0
javalang>=0.13.0
pandas>=2.0.0
python-dotenv>=1.0.0
GitPython>=3.1.0
